+++ wechat/src/main/scala/zhongl/stream/oauth2/wechat/WeWork.scala
@@ -1,12 +1,9 @@
 package zhongl.stream.oauth2.wechat
 
 import java.net.URLEncoder
-import java.util.Base64
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.HttpEntity.Strict
-import akka.http.scaladsl.model.StatusCodes.OK
 import akka.http.scaladsl.model.Uri.Query
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.headers.Location
@@ -77,19 +74,6 @@ class WeWork(authenticated: (UserInfo, Uri) => HttpResponse)(implicit system: Ac
 
 object WeWork extends JsonSupport {
 
-  final object Content {
-    def unapply(result: HttpResponse): Option[ByteString] = result match {
-      case HttpResponse(OK, _, Strict(ContentTypes.`application/json`, content), _) => Some(content)
-      case _                                                                        => None
-    }
-  }
-
-  final def complainIllegalResponse[T](pf: PartialFunction[HttpResponse, T])(implicit mat: Materializer): HttpResponse => T = {
-    case r if pf.isDefinedAt(r)                      => pf(r)
-    case HttpResponse(status, _, Strict(_, data), _) => throw new IllegalStateException(s"$status - ${data.utf8String}")
-    case HttpResponse(status, _, entity, _)          => entity.discardBytes(); throw new IllegalStateException(s"$status - ignored chunk")
-  }
-
   object AccessTokenE {
     def unapply(arg: ByteString): Option[AccessToken] = arg.as[AccessToken].toOption
   }
@@ -106,11 +90,6 @@ object WeWork extends JsonSupport {
     def unapply(arg: ByteString): Option[Err] = arg.as[Err].toOption
   }
 
-  @inline
-  final def base64Decode(v: String): String = new String(Base64.getUrlDecoder.decode(v))
-
-  @inline
-  final def base64Encode(v: String): String = Base64.getUrlEncoder.encodeToString(v.getBytes)
 
   def apply(authenticated: (UserInfo, Uri) => HttpResponse)(implicit system: ActorSystem): WeWork = new WeWork(authenticated)
 }
